inline one-line method ErrorMessages.screenshot()
had to move tests from ErrorMessagesTest to ScreenShotLaboratoryTest (which is the appropriate place)
